+++ service-worker.js
@@ -27,8 +27,8 @@ importScripts('lib/util/misc/Wikify.js');
 // Based on https://deanhume.com/displaying-a-new-version-available-progressive-web-app/
 
 const info = {
-    version: '0.9.99',
-    release: 151
+    version: '0.9.991',
+    release: 1
 };
 
 const cacheName = 'v' + info.version + "r" + info.release;